requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
groq==0.34.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.2.17
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-text-splitters==0.2.4
langsmith==0.1.147
loguru==0.7.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.4
packaging==23.2
pandas==2.3.3
pillow==10.4.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==22.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
streamlit==1.28.0
tenacity==8.5.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
